requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
python-dotenv==1.0.1
Werkzeug==3.0.1
zipp==3.17.0